% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_positions.R
\name{check_positions}
\alias{check_positions}
\title{Check Positions}
\usage{
check_positions(ptbl, P)
}
\arguments{
\item{ptbl}{tbd}

\item{P}{tbd}
}
\value{
tbd
}
\description{
Check Positions
}
\examples{
"tbd"
}
