% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethwbtc_net_liquidity.R
\docType{data}
\name{ethwbtc_net_liquidity}
\alias{ethwbtc_net_liquidity}
\title{ETH-WBTC 0.3\% Net Liquidity}
\format{
A data frame with 37,478 rows and 4 variables:
\describe{
  \item{block_number}{Ethereum mainnet block number}
  \item{tick_lower}{The lower price range for a Uniswap v3 position(s)}
  \item{tick_upper}{The higher price range for a Uniswap v3 position(s)}
  \item{net_adjusted_liquidity}{The total liquidity (adjusted by 1e-13) available between the ticks as of the block_number (i.e. cumulative of deposits and withdraws across all positions with the same tick_lower-tick_upper)}
  \item{liquidity}{The raw liquidity (adjusted * 1e13) available between the ticks as of the block_number. Required for correct swap sizing.}
}
}
\source{
\href{https://flipsidecrypto.xyz/}{Flipside Crypto}
}
\usage{
ethwbtc_net_liquidity
}
\description{
Net Liquidity between ticks for the Ethereum mainnet ETH-WBTC 0.3\% Uniswap v3 Liquidity Pool
as of block height from inception to block 17,000,000. See overview.Rmd vignette for exact Flipside Crypto query.
}
\keyword{datasets}
