% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_recalculation_price.R
\name{find_recalculation_price}
\alias{find_recalculation_price}
\title{Find Recalculation Price}
\usage{
find_recalculation_price(
  ptbl,
  P,
  price_up = TRUE,
  decimal_adjustment = 1,
  yx = TRUE
)
}
\arguments{
\item{ptbl}{Liquidity Positions table of the form tick_lower, tick_upper, liquidity. See ?ethwbtc_net_liquidity for an example.
This included dataset is net_liquidity, so it should be filtered to the max blockheight for a given tick_lower-tick_upper.}

\item{P}{Specific price in human readable format, converted to tick for flagging using ?get_closest_tick with tick_spacing = 1.}

\item{price_up}{Is the price rising (TRUE) or falling (FALSE) given specified format for `P`. Default TRUE.}

\item{decimal_adjustment}{The difference in the tokens decimals, e.g., 1e10 for ETH vs BTC. 1e12 for USDC vs ETH; 1 (default) for most ERC20/ETH.}

\item{yx}{Whether price is already in Token 1 / Token 0 format or inverted. ETH per USDC is how the ETH Mainnet 0.05\% pool works but is not friendly for human interpretation. Default is TRUE.}
}
\value{
tbd
}
\description{
Helper function for identifying the next tick where liquidity must be recalculated (i.e., at least 1 position enters or exits being active).
Note, pass both active and inactive positions with liquidity, see ?check_positions.
}
\examples{
ptbl <- ethwbtc_net_liquidity
# filter to a small subset for easier manual review
ptbl <- ptbl[(ptbl$tick_lower < 100000 & ptbl$tick_lower > 0), ]
# append price conversions for easier manual review
ptbl$price_lower <- tick_to_price(ptbl$tick_lower, 1e10, TRUE)
ptbl$price_upper <- tick_to_price(ptbl$tick_upper, 1e10, TRUE)
# 39.36252 ETH/BTC current price in correct yx format
P = 39.36252
tick = 267180
# Check for next higher price. 1e10 decimal_adjustment.
# Liquidity needs to be recalculated at 40.07743 ETH / BTC (the 92100-267180 position falls out of range!)
find_recalculation_price(ptbl, P, price_up = TRUE, 1e10, yx = TRUE)
}
