% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_position_balance.R
\name{get_position_balance}
\alias{get_position_balance}
\title{Title}
\usage{
get_position_balance(
  position_L,
  sqrtpx96,
  t_lower,
  t_upper,
  decimal_x = 1e+18,
  decimal_y = 1e+18
)
}
\arguments{
\item{position_L}{The marginal liquidity provided by a position when created and the net of any increases or decreases
in liquidity of that position afterwards. This is usable on `ethwbtc_net_liquidity` but it would be the balance of all positions in a range.}

\item{sqrtpx96}{current price in uint160 format.See ?price_to_sqrtpx96 or read a pool contract's sqrtPriceX96 within it's slot0 on etherscan to get this value.}

\item{decimal_x}{The decimals used in token 0, e.g., 1e6 for USDC, 1e8 for WBTC.}

\item{decimal_y}{The decimals used in token 1, e.g., 1e18 for WETH.}

\item{tick_lower}{The low tick in a liquidity position, see ?get_closest_tick to convert a price to a tick.}

\item{tick_upper}{The upper tick in a liquidity position, see ?get_closest_tick to convert a price to a tick.}
}
\value{
a list of `token0` balance and `token1` balance adjusted for the provided decimals for a given position at any price
}
\description{
Title
}
\examples{
# This is real world MKR/LINK position on Optimism. In the range of 0.00667 - 0.02 MKR/LINK
given 343255264548669212 liquidity, at the price 0.00928 MKR/LINK price, there are
 1.136317 LINK and 0.005027558 MKR in the position.
tick_lower = -50100
tick_upper = -39120
L = as.bigz('343255264548669212')
sqrtpx96 = as.bigz('7632249339194475209177795127')
get_position_balance(position_L = L, sqrtpx96 = sqrtpx96,
tick_lower = tick_lower, tick_upper = tick_upper,
decimal_x = 1e18, decimal_y = 1e18)
}
