% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquidity_asof_block.R
\name{liquidity_asof_block}
\alias{liquidity_asof_block}
\title{Liquidity as of Block}
\usage{
liquidity_asof_block(ptbl, blockheight, clean = TRUE)
}
\arguments{
\item{ptbl}{Liquidity Positions table of the form tick_lower, tick_upper, liquidity. See ?ethwbtc_net_liquidity for an example.
This included dataset is net_liquidity, this function filters the data to a blockheight for a given tick_lower-tick_upper.}

\item{blockheight}{The block number to partition and filter to.}

\item{clean}{remove erroneous liquidity (0 or occasionally negative in case of data issue with a pool).}
}
\value{
A subset of `ptbl` with 1 row per tick_lower-tick_upper.
}
\description{
Filters a net liquidity positions table (see ?ethwbtc_net_liquidity for an example) to liquidity across ranges as of a block
by partitioning tick_lower-tick_upper and selecting the maximum block <= blockheight.
}
\examples{
data(ethwbtc_net_liquidity)
l1600 = liquidity_asof_block(ethwbtc_net_liquidity, 16000000)
# all tick_lower-tick_upper combos aggregated as of block 16M
max(table(l1600$tick_lower, l1600$tick_upper)) == 1
}
