% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tokens_to_range.R
\name{match_tokens_to_range}
\alias{match_tokens_to_range}
\title{Title}
\usage{
match_tokens_to_range(x = NULL, y = NULL, P, pa, pb, yx = TRUE)
}
\arguments{
\item{x}{number of token 0, e.g., WBTC in ETH-WBTC 0.3 percent pool on ETH Mainnet. Must be NULL if y is not NULL.}

\item{y}{number of token 1, e.g., ETH in ETH-WBTC 0.3 percent pool on ETH Mainnet. Must be NULL if x is not NULL.}

\item{P}{Current price of the pool, be aware of which unit of account you are using.}

\item{pa}{The minimum price in position.}

\item{pb}{The maximum price in position.}

\item{yx}{Whether current price P is in Token 1 / Token 0 (y/x) format already. Default TRUE.}
}
\value{
A list of amount_x, amount_y (i.e., filling in the NULL), current_price P, min_price pa, max_price pb.
}
\description{
Given current price, a price range, and an amount of 1 token; identify how much of the other token is required create
the Uni v3 position.
}
\examples{
match_tokens_to_range(x = NULL, # how many wrapped bitcoin, WBTC
y = 12.549, # to match to 12.549 ETH
P = 0.1, # given the current price is 0.1 BTC / ETH
pa = 0.05, # minimum price is 0.05 BTC / ETH (ETH is weaker than current)
pb = 0.25, # maximum price is 0.25 BTC / ETH (ETH is stronger than current)
yx = FALSE) # the unit of account used: BTC/ETH is NOT the pool's Token 1 / Token 0 price.
Function will handle all inversions.

match_tokens_to_range(x = NULL, # how many wrapped bitcoin WBTC
y = 12.549, # to match to 12.549 ETH
P = 10, # given the current price is 10 ETH/BTC
pa = 4, # minimum price is 4 ETH / BTC (BTC is weaker than current), note: this is 0.25^-1
pb = 20, # maximum price is 20 ETH / BTC (BTC is stronger than current), note: this is 0.05^-1
yx = TRUE) # ETH/BTC IS Token 1 / Token 0 (y/x) price.
}
