% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_all_tokens.R
\name{price_all_tokens}
\alias{price_all_tokens}
\title{Price All Tokens}
\usage{
price_all_tokens(x, y, P, pa = NULL, pb = NULL, yx = TRUE)
}
\arguments{
\item{x}{number of token 0, e.g., WBTC in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{y}{number of token 1, e.g., ETH in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{P}{Current (human readable) price of the pool, see ?sqrtpx96_to_price if using slot0 from a pool contract.}

\item{pa}{The minimum price in position, Must be NULL if pb is not NULL.}

\item{pb}{The maximum price in position. Must be NULL if pa is not NULL.}

\item{yx}{Whether current price P is in Token 1 / Token 0 (y/x) format already. Default TRUE.}
}
\value{
A list of amount_x, amount_y, current_price P, min_price pa, max_price pb (i.e., filling in the NULL).
}
\description{
Given an amount of tokens, a current price, and either the min or max price of a range; identify the corresponding
max or min price for the Uni v3 position to use all tokens x and y. NOTE: not all pools allow all prices, see:
?get_closest_tick for more info.
}
\examples{
price_all_tokens(x = 100, # given 100 BTC
y = 1000, # and 1000 ETH
P = 0.1, # and a current price of 0.1 BTC/ETH
 pa = NULL, # what minimum BTC/ETH price
 pb = 0.25, # matches 0.25 BTC/ETH max price
 yx = FALSE) # the unit of account used: BTC/ETH is NOT the pool's Token 1 / Token 0 price.
 Function will handle all inversions.

price_all_tokens(x = 100, # given 100 BTC
y = 1000, # 1000 ETH
P = 10, # a current price of 10 ETH/BTC
 pa = 4, # a minimum price of 4 ETH/BTC, note: this is 0.25^-1
 pb = NULL, # what maximum ETH/BTC price uses all tokens.
 yx = TRUE) # the unit of account used: ETH/BTC IS the pool's Token 1 / Token 0 price.
}
