% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_across_ticks.R
\name{swap_across_ticks}
\alias{swap_across_ticks}
\title{Swap Across Ticks}
\usage{
swap_across_ticks(
  ptbl,
  sqrtpx96,
  fee_tbl = NULL,
  trade_record = NULL,
  dx = NULL,
  dy = NULL,
  decimal_x = 1e+18,
  decimal_y = 1e+18,
  fee = 0.003
)
}
\arguments{
\item{ptbl}{tbd}

\item{sqrtpx96}{tbd}

\item{fee_tbl}{tbd}

\item{trade_record}{tbd}

\item{dx}{tbd}

\item{dy}{tbd}

\item{decimal_x}{tbd}

\item{decimal_y}{tbd}

\item{fee}{tbd}
}
\value{
tbd
}
\description{
Swap Across Ticks
}
\examples{
"tbd"
}
