% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_recalculation_price.R
\name{find_recalculation_price}
\alias{find_recalculation_price}
\title{Find Recalculation Price}
\usage{
find_recalculation_price(positions, current_price, price_up = TRUE)
}
\arguments{
\item{positions}{tbd}

\item{current_price}{tbd}

\item{price_up}{tbd}
}
\value{
tbd
}
\description{
Find Recalculation Price
}
\examples{
"tbd"
}
