% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_tick.R
\name{get_closest_tick}
\alias{get_closest_tick}
\title{Get Closest Tick}
\usage{
get_closest_tick(
  desired_price,
  tick_spacing = 60,
  decimal_adjustment = 1,
  yx = TRUE
)
}
\arguments{
\item{desired_price}{Your desired price, note it is important to know the unit of account. 0.05 BTC/ETH is 20 ETH/BTC.}

\item{tick_spacing}{The pool's minimum tick spacing, default 0.3 percent pool, i.e., 60.}

\item{decimal_adjustment}{The difference in the tokens decimals, e.g., 1e10 for ETH vs BTC. 1e12 for USDC vs ETH; 1 (default) for most ERC20/ETH.}

\item{yx}{Whether price is already in Token 1 / Token 0 format or inverted. ETH per USDC may be how the pool functions but is not friendly for human interpretation. Default is TRUE.}
}
\value{
a list of `desired_price` (the input), the closest allowable price `actual_price``, and the `tick` of that allowable price.
}
\description{
Depending on the Uniswap v3 Pool fee tier, only specific ticks are allowed to be
used in positions. In 0.05 percent pools, the tick spacing is 10 minimum. In 0.3 percent, the minimum is 60 ticks.
This function identifies the closest allowed tick given tick spacing.
}
\examples{
# 0.05 BTC / ETH is NOT Y/X accounting for the pool, so yx = FALSE.
get_closest_tick(0.05,
tick_spacing = 60, # 0.3 percent fee pool
decimal_adjustment = 1e10, # ETH 18 decimals vs BTC 8 decimals.
 yx = FALSE)

# closest allowable tick: -260220
get_closest_tick(20, tick_spacing = 60, decimal_adjustment = 1e10, yx = TRUE)

# You can use tick_spacing = 1 to get the exact reverse of ?tick_to_price but it may not be allowed for making a new position.
get_closest_tick(desired_price = 0.05004423, tick_spacing = 1, decimal_adjustment = 1e10, yx = FALSE) # tick -260220
}
