% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liquidity.R
\name{get_liquidity}
\alias{get_liquidity}
\title{Calculate Liquidity}
\usage{
get_liquidity(
  x,
  y,
  sqrtpx96,
  decimal_x = 1e+18,
  decimal_y = 1e+18,
  tick_lower,
  tick_upper
)
}
\arguments{
\item{x}{number of token 0, e.g., WBTC in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{y}{number of token 1, e.g., ETH in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{sqrtpx96}{current price in uint160 format.See ?price_to_sqrtpx96 or read a pool contract's sqrtPriceX96 within it's slot0 on etherscan to get this value.}

\item{decimal_x}{The decimals used in token 0, e.g., 1e6 for USDC, 1e8 for WBTC.}

\item{decimal_y}{The decimals used in token 1, e.g., 1e18 for WETH.}

\item{tick_lower}{The low tick in a liquidity position, see ?get_closest_tick to convert a price to a tick.}

\item{tick_upper}{The upper tick in a liquidity position, see ?get_closest_tick to convert a price to a tick.}
}
\value{
Returns a single Big Integer value, the liquidity contributed by the position.
}
\description{
This function calculates the liquidity provided by a range using its amount of tokens (adjusting for their decimals),
the current price in sqrtpx96 format (see ?price_to_sqrtpx96), and the range (tick_lower, tick_upper)
}
\examples{
# See: https://etherscan.io/tx/0xc10cddba3df56e6fba4f9a88b132cc9d4440ff31bb0c4926dc9d9ca652faf376#eventlog
# In Block 12,376,757: 1 BTC and 16.117809469 ETH were added to pool with a range of 257760 to 258900
# the price at the time was 16.52921 ETH / BTC (sqrtpx96 = 32211102662183904786754519772954624)
# it resulted in NFT Position 1005 with liquidity 1429022393248418
# Within 0.0001\% (some precision loss expected)
get_liquidity(x = 1, y = 16.117809469,
sqrtpx96 = gmp::as.bigz('32211102662183904786754519772954624'),
decimal_x = 1e8, decimal_y = 1e18,
tick_lower = 257760,
tick_upper = 258900) / gmp::as.bigz('1429022393248418') - 1 < 0.000001
}
