% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_all_tokens.R
\name{price_all_tokens}
\alias{price_all_tokens}
\title{Price All Tokens}
\usage{
price_all_tokens(
  x,
  y,
  sqrtpx96,
  decimal_x = 1e+18,
  decimal_y = 1e+18,
  tick_lower,
  tick_upper
)
}
\arguments{
\item{x}{number of token 0, e.g., WBTC in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{y}{number of token 1, e.g., ETH in ETH-WBTC 0.3 percent pool on ETH Mainnet.}

\item{sqrtpx96}{current price in uint160 format.See ?price_to_sqrtpx96 or read a pool contract's sqrtPriceX96 within it's slot0 on etherscan to get this value.}

\item{decimal_x}{The decimals used in token 0, e.g., 1e6 for USDC, 1e8 for WBTC.}

\item{decimal_y}{The decimals used in token 1, e.g., 1e18 for WETH.}

\item{tick_lower}{The low tick in a liquidity position, see ?get_closest_tick to convert a price to a tick. Otherwise NULL if tick_upper is provided.}

\item{tick_upper}{Default NULL. Otherwise the upper tick in a liquidity position, see ?get_closest_tick to convert a price to a tick.}
}
\value{
A list of `amount_x`, `amount_y`, `sqrtpx96`, `P`, `tick_lower`, `tick_upper`, `price_lower`, `price_upper`
where tick_lower or tick_upper is replaced with its matched value and their corollary human readable prices are in Token 1 / Token 0 (y/x) format.
}
\description{
Given an amount of tokens, a current price, and either the min or max price of a range; identify the corresponding
max or min price for the Uni v3 position to use all tokens x and y. NOTE: not all pools allow all prices, see:
?get_closest_tick for more info.
}
\examples{
# Matches ?match_tokens_to_range example for ETH/WBTC pool.
price_all_tokens(x = 1,
                 y = 16.11781,
                sqrtpx96 = '32211102662183904786754519772954624',
                decimal_x = 1e8,
                decimal_y = 1e18,
                tick_lower = NULL,
                tick_upper = 258900)$tick_lower == 257760
}
