% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_to_sqrtpx96.R
\name{price_to_sqrtpx96}
\alias{price_to_sqrtpx96}
\title{Price to SQRTPX96}
\usage{
price_to_sqrtpx96(P)
}
\arguments{
\item{P}{price in standard form (e.g., 0.05 BTC/ETH). Be aware of the pool specific
token 0 and token 1 as this function takes your number as is.}
}
\value{
Big Integer price in sqrtpx96 format.
}
\description{
Uniswap stores prices as square roots in 64.96 (64 bits integer, 96 bit fractional)
assume sqrt price is a rational and use gmp big integer
}
\examples{
price_to_sqrtpx96(0.05) == gmp::as.bigq('17715955711429570570949230592')
}
