% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_to_price.R
\name{tick_to_price}
\alias{tick_to_price}
\title{Tick to Price}
\usage{
tick_to_price(tick, decimal_adjustment = 1, yx = TRUE)
}
\arguments{
\item{tick}{The numeric tick, e.g., -255410.}

\item{decimal_adjustment}{The difference in the tokens decimals, e.g., 1e10 for ETH vs BTC. 1e12
for USDC vs ETH.}

\item{yx}{Whether tick is already in Token 1 / Token 0 format or inverted. ETH per USDC may be
how the pool functions but is not friendly for human interpretation. Default is TRUE.}
}
\value{
A numeric price in desired format. See ?get_closest_tick for the reverse process.
}
\description{
Converts a Uniswap v3 tick to a human readable price accounting
for differences in the token decimals and whether price is desired in
Token 1 / Token 0 (y/x) or inverted (x/y) format. See ?get_closest_tick for the reverse of this function.
}
\examples{
 # 1,351.327 USDC per ETH; 6 decimals vs 18 decimals -> 1e12 adjustment.
 tick_to_price(204232, decimal_adjustment = 1e12, yx = TRUE)

 # You can multiply tick by -1 to get the equivalent of inverting prices.
 tick_to_price(tick = -204232, decimal_adjustment = 1e12, yx = FALSE)

 # 0.05004423 WBTC per ETH; 8 decimals vs 18 decimals -> 1e10 adjustment.
tick_to_price(-260220, decimal_adjustment = 1e10, yx = FALSE)
}
